[KFKESQ-11] Started refactoring of the producer DSL
