+++ kafkaesque-core/src/test/java/in/rcard/kafkaesque/KafkaesqueProducerTest.java
@@ -4,6 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.BDDMockito.given;
 
+import in.rcard.kafkaesque.producer.KafkaesqueProducer;
 import java.time.Duration;
 import java.time.temporal.ChronoUnit;
 import java.util.Arrays;
